requirements.txt
maturin
setuptools
wheel

numpy
numba
PyFixedReps-andnp
